- Moves fashion story template to 'fashion-on-the-go'
