+++ packages/stickers/src/fashion-on-the-go/arrowDark.js
