+++ packages/stickers/src/fashion-on-the-go/arrowLight.js
