+++ packages/stickers/src/fashion-on-the-go/index.js
