+++ packages/stickers/src/fashion-on-the-go/starburst.js
