+++ packages/stickers/src/fashion-on-the-go/starburstSolid.js
