+++ packages/stickers/src/fashion-on-the-go/starburstWithArrow.js
