+++ packages/stickers/src/index.js
@@ -20,7 +20,7 @@ import * as beautyStickers from './fresh-and-bright';
 import * as cookingStickers from './food-and-stuff';
 import * as doersGetMoreDoneStickers from './doers-get-more-done';
 import * as weeklyEntertainmentStickers from './weekly-entertainment';
-import * as fashionStickers from './fashion';
+import * as fashionOnTheGoStickers from './fashion-on-the-go';
 import * as fitnessStickers from './fitness';
 import * as travelStickers from './travel';
 import * as wellbeingStickers from './wellbeing';
@@ -30,7 +30,7 @@ export default {
   ...cookingStickers,
   ...doersGetMoreDoneStickers,
   ...weeklyEntertainmentStickers,
-  ...fashionStickers,
+  ...fashionOnTheGoStickers,
   ...fitnessStickers,
   ...travelStickers,
   ...wellbeingStickers,
