+++ packages/templates/src/raw/fashion-on-the-go/template.json
